debug for smart smoothing
adds a debug that shows what the dterm is multiplied by. It shows it multiplied by 1000 though, so if smart smoothing shrinks your dterm by half it will show 500. If it leaves the dterm untouched it will show 1000. Just a way to make this easier to see in a bbl.