remove rate limit make it static
rate limit was never never used. It just allowed the user to set the max rate of their rates. But you already do that when you set up your rates, so its was super pointless. just lock the max setpoint to 1998 which is what this was always defaulted to anyhow.