Allow Dterm Boosting
as error is decreasing dterm will be boosted, while error is increasing dterm will be shrunk. The amount dterm is boosted or shrunk depends on how quickly error is increasing/decreasing. The faster it is changing the more we will boost/shrink d.  As error transitions from growing to shrinking the dterm will go from being shrunk, back to normal strength, and then to boosted, then back to normal strength. Gain scheduling FTW :)